+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+### Next version
+
+### ✨ Improved
+
+* Break out of `Gort.observe()`` on keyboard interrupt
+
+
 ## 0.8.0 - January 13, 2024
 
 ### 🚀 New